+++ src/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.0"
+__version__ = "1.2.0b1"
 
 __all__ = ["__version__"]