+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [1.2.0b1] - 2026-07-01
+
 ### Added
 
 - **`price-scout db migrate check`**: Exit-code based migration check for automation (Docker, CI)